requirements.txt
flask
python-dotenv
openai
requests
pandas
matplotlib